+++ CHANGES.md
@@ -8,6 +8,7 @@ New:
 Fixed:
 
 - Fixed `source.drop` and `source.dump` clock initialization.
+- Fixed bogus report of non-monotonous PTS content when using raw ffmpeg content.
 
 ---